refactor(domain): 패키지 이름 변경
model 패키지 이름 domain으로 변경
