+++ src/main/java/domain/Cars.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import java.util.Collections;
 import java.util.List;
