+++ src/main/java/domain/MovesGenerator.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 public class MovesGenerator {
